Factor out election
